Improvements to the node navigation
Split navigator and state. Make the node model and navigator stateless.
